educational-ai-interpretability/requirements.txt
openai>=1.3.0
textstat>=0.7.2
nltk>=3.8.1
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
jupyter>=1.0.0
python-dotenv>=1.0.0